+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - Logging when Crestron input cards, output cards and streamers fail to instantiate
+
 ## [13.4.1] - 2019-04-19
 ### Added
  - Added RoutingCache methods for source/destination state lookup